Decrease CONTROL.rainingSafeRounds to 9
This will allow abilities and damage to still be used against x90 bosses. Right now, getting through those bosses is very slow, and gives trolls an opportunity to spam wormholes. If they deploy 11 wormholes, we miss the next x00 boss. This seems to be happening on a regular basis.

We don't know if this will actually make things better: it may be that, by not pausing on the x90 levels, we end up dealing too much damage and overflowing damage onto the x00 bosses. But I think it is worth a shot.